Added `examples` helper
